+++ src/main/java/org/springframework/beans/factory/Aware.java
@@ -1,0 +1,10 @@
+package org.springframework.beans.factory;
+
+/**
+ * 标记类接口，实现该接口可以被Spring容器感知
+ *
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public interface Aware {}