+++ src/test/java/org/springframework/beans/factory/HelloWorldService.java
@@ -1,0 +1,51 @@
+package org.springframework.beans.factory;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.context.ApplicationContext;
+import org.springframework.beans.context.ApplicationContextAware;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/10/8
+ * @project: sbs-small-spring
+ */
+public class HelloWorldService
+    implements BeanNameAware, BeanClassLoaderAware, ApplicationContextAware, BeanFactoryAware {
+
+  private ApplicationContext applicationContext;
+
+  private BeanFactory beanFactory;
+
+  public String sayHelloWorld() {
+    System.out.println("hello world");
+    return "hello world";
+  }
+
+  @Override
+  public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
+    this.beanFactory = beanFactory;
+  }
+
+  @Override
+  public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+    this.applicationContext = applicationContext;
+  }
+
+  public ApplicationContext getApplicationContext() {
+    return applicationContext;
+  }
+
+  public BeanFactory getBeanFactory() {
+    return beanFactory;
+  }
+
+  @Override
+  public void setBeanClassLoader(ClassLoader classLoader) {
+    System.out.println("ClassLoader：" + classLoader);
+  }
+
+  @Override
+  public void setBeanName(String name) {
+    System.out.println("Bean Name is：" + name);
+  }
+}